Change position of dark mode toggle
