+++ site/src/layout/Footer/Footer.js
@@ -1,10 +1,13 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useContext } from 'react';
 
 import { SocialIcon } from '@components';
+import { Theme } from '@style/theme';
 
 import * as S from './style';
 
 export const Footer = () => {
+  const { styledTheme, theme, toggleTheme } = useContext(Theme.Context);
+
   return (
     <Fragment>
       <S.Border />
@@ -19,13 +22,26 @@ export const Footer = () => {
           <S.ListItem>
             <SocialIcon type="spotify" />
           </S.ListItem>
-          <S.ListItem>
+          {/* <S.ListItem>
             <SocialIcon type="twitch" />
-          </S.ListItem>
+          </S.ListItem> */}
           <S.ListItem>
             <SocialIcon type="twitter" />
           </S.ListItem>
         </S.List>
+
+        <S.Toggle
+          aria-label="Toggle dark mode on or off"
+          checked={theme === 'dark'}
+          onChange={toggleTheme}
+          checkedIcon={<S.Sun />}
+          uncheckedIcon={<S.Moon />}
+          onColor={styledTheme.text}
+          onHandleColor={styledTheme.body}
+          offColor={styledTheme.text}
+          offHandleColor={styledTheme.body}
+          handleDiameter={20}
+        />
       </S.Footer>
     </Fragment>
   );
