+++ site/src/layout/index.js
@@ -1,32 +1,15 @@
 /* eslint-disable react/jsx-pascal-case */
-import React, { Fragment, useContext } from 'react';
+import React, { Fragment } from 'react';
 
 import { Seo } from './Seo/Seo';
 import { Nav } from './Nav/Nav';
 import { Footer } from './Footer/Footer';
 
-import { Theme } from '@style/theme';
-
 import * as S from './style';
 
 export const Layout = ({ children, path }) => {
-  const { styledTheme, theme, toggleTheme } = useContext(Theme.Context);
-
   return (
     <Fragment>
-      <S.Toggle
-        aria-label="Toggle dark mode on or off"
-        checked={theme === 'dark'}
-        onChange={toggleTheme}
-        checkedIcon={<S.Sun />}
-        uncheckedIcon={<S.Moon />}
-        onColor={styledTheme.text}
-        onHandleColor={styledTheme.body}
-        offColor={styledTheme.text}
-        offHandleColor={styledTheme.body}
-        handleDiameter={20}
-      />
-
       <S.PageWrapper>
         <Seo path={path} title="Chris Carrick - Developer" />
         <Nav />
